+++ test/electron/fixtures/app/configs/feature-flags-on-send-error.js
@@ -1,0 +1,4 @@
+const featureFlagsConfig = require('./feature-flags')
+const onSendErrorConfig = require('./on-send-error')
+
+module.exports = () => Object.assign({}, featureFlagsConfig(), onSendErrorConfig())