+++ test/electron/fixtures/app/configs/on-send-error.js
@@ -1,0 +1,9 @@
+module.exports = () => {
+  return {
+    onSendError (event) {
+      event.context = 'checkout page'
+      event.addMetadata('account', { type: 'VIP', verified: true })
+      event.addMetadata('account', 'status', "it's complicated")
+    }
+  }
+}